Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3212 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/ui/headerwindow.cpp
@@ -857,7 +857,7 @@ bool qm::EditHeaderItem::canSelectAll()
 
 unsigned int qm::EditHeaderItem::getLineCount() const
 {
-	return Window(getHandle()).sendMessage(EM_GETLINECOUNT);
+	return static_cast<unsigned int>(Window(getHandle()).sendMessage(EM_GETLINECOUNT));
 }
 
 unsigned int qm::EditHeaderItem::parseMultiline(const WCHAR* pwszMultiline)
